final version with validate

+++ charity/src/main/java/com/tuyet/charity/controller/CommentController.java
@@ -9,6 +9,7 @@ import com.tuyet.charity.service.NotificationService;
 import com.tuyet.charity.service.PostService;
 import com.tuyet.charity.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
@@ -18,10 +19,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @RestController
 public class CommentController {
@@ -38,7 +36,10 @@ public class CommentController {
     private NotificationService notificationService;
 
     @Autowired
-    private Notification notification;
+    private ApplicationContext applicationContext;
+
+//    @Autowired
+//    private Notification notification;
 
 //    @GetMapping("/comments")
 //    public List<Comment> getAllCommentsPost(@RequestParam(value = "post-id") Integer postId){
@@ -62,30 +63,63 @@ public class CommentController {
 
     @PostMapping("comments/")
     public ResponseEntity<Object> createComment(@RequestParam(value = "post-id") Integer postId,
-                                                @Valid @RequestBody Comment comment, OAuth2Authentication auth){
-        Post createdPost = postService.getPost(postId);
-        User owner = userDetailsService.getUserByUsername(auth.getName());
-        //tao comment
-        comment.setPost(createdPost);
-        comment.setUser(owner);
-        Comment createdComment=commentService.createComment(comment);
-
-        //tao notification
-        if(!auth.getName().equals(createdPost.getOwnerPost().getUsername())){
-            notification.setPost(createdPost);
-            notification.setUser(createdPost.getOwnerPost());
-            notification.setType(2);
-            notificationService.createNotification(notification);
+                                                @Valid @RequestBody Comment comment, OAuth2Authentication auth,
+                                                BindingResult result){
+        if(result.hasErrors()){
+            Map<String, String> errors = new HashMap<>();
+            result.getAllErrors().forEach((error) -> {
+                String fieldName = ((FieldError) error).getField();
+                String errorMessage = error.getDefaultMessage();
+                errors.put(fieldName, errorMessage);
+            });
+            return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
         }
 
-        return new ResponseEntity<>(createdComment, HttpStatus.CREATED);
+        try {
+            Post createdPost = postService.getPost(postId);
+            User owner = userDetailsService.getUserByUsername(auth.getName());
+            //tao comment
+            comment.setPost(createdPost);
+            comment.setUser(owner);
+            Comment createdComment = commentService.createComment(comment);
+
+            Notification notification = applicationContext.getBean(Notification.class);
+            //tao notification
+            if (!auth.getName().equals(createdPost.getOwnerPost().getUsername())) {
+                notification.setPost(createdPost);
+                notification.setUser(createdPost.getOwnerPost());
+                notification.setType(2);
+                System.out.println(notification.getType());
+                notificationService.createNotification(notification);
+            }
+
+            return new ResponseEntity<>(createdComment, HttpStatus.CREATED);
+        } catch (NoSuchElementException ex){
+            Map<String, String> msg = new HashMap<>();
+            msg.put("error","This comment does not exist");
+            return new ResponseEntity<>(msg, HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/comments/{commentId}")
-    public void deleteComment(@PathVariable Integer commentId){
-        //bat comment khong ton tai
-        Comment createdComment = commentService.getCommentById(commentId);
-        commentService.deleteComment(createdComment);
+    public ResponseEntity<Object> deleteComment(@PathVariable Integer commentId, OAuth2Authentication auth){
+        try {
+            Comment comment = commentService.getCommentById(commentId);
+            //kt owner comment
+            if(!auth.getName().equals(comment.getUser().getUsername())){
+                Map<String, String> msg = new HashMap<>();
+                msg.put("error","This user has no permission");
+                return new ResponseEntity<>(msg, HttpStatus.BAD_REQUEST);
+            }
+
+            Comment createdComment = commentService.getCommentById(commentId);
+            commentService.deleteComment(createdComment);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } catch (NoSuchElementException ex){
+            Map<String, String> msg = new HashMap<>();
+            msg.put("error","This comment does not exist");
+            return new ResponseEntity<>(msg, HttpStatus.NOT_FOUND);
+        }
     }
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
